+++ guillotina/db/storages/pg.py
@@ -235,6 +235,18 @@ RETURNING id;
 ''')
 
 
+def restart_conn_on_exception(f):
+    async def decorator(self: 'PostgresqlStorage', *args, **kwargs):
+        try:
+            return await f(self, *args, **kwargs)
+        except (asyncpg.exceptions.PostgresConnectionError,
+                asyncpg.exceptions.InterfaceError) as ex:
+            await self._check_bad_connection(ex)
+            raise
+
+    return decorator
+
+
 # how long to wait before trying to recover bad connections
 BAD_CONNECTION_RESTART_DELAY = 0.25
 
@@ -646,13 +658,10 @@ WHERE tc.constraint_name = '{}_parent_id_id_key' AND tc.constraint_type = 'UNIQU
             await conn.execute("DROP TABLE IF EXISTS {};".format(self._blobs_table_name))
             await conn.execute("DROP TABLE IF EXISTS {};".format(self._objects_table_name))
 
+    @restart_conn_on_exception
     async def open(self):
-        try:
-            conn = await self.pool.acquire(timeout=self._conn_acquire_timeout)
-            return conn
-        except asyncpg.exceptions.InterfaceError as ex:
-            async with self.lock:
-                await self._check_bad_connection(ex)
+        conn = await self.pool.acquire(timeout=self._conn_acquire_timeout)
+        return conn
 
     async def close(self, con):
         try:
@@ -755,26 +764,22 @@ WHERE tc.constraint_name = '{}_parent_id_id_key' AND tc.constraint_type = 'UNIQU
         txn.add_after_commit_hook(self._txn_oid_commit_hook, oid)
 
     async def _check_bad_connection(self, ex):
-        if str(ex) in ('cannot perform operation: connection is closed',
-                       'connection is closed', 'pool is closed'):
-            if (time.time() - self._connection_initialized_on) > BAD_CONNECTION_RESTART_DELAY:
-                # we need to make sure we aren't calling this over and over again
-                return await self.restart_connection()
-
-    async def get_next_tid(self, txn):
         async with self.lock:
             # we do not use transaction lock here but a storage lock because
             # a storage object has a shard conn for reads
-            try:
-                return await self._stmt_next_tid.fetchval()
-            except asyncpg.exceptions.InterfaceError as ex:
-                await self._check_bad_connection(ex)
-                raise
+            for err in ('connection is closed', 'pool is closed', 'connection was closed'):
+                if err in str(ex):
+                    if (time.time() - self._connection_initialized_on) > BAD_CONNECTION_RESTART_DELAY:
+                        # we need to make sure we aren't calling this over and over again
+                        return await self.restart_connection()
 
+    @restart_conn_on_exception
+    async def get_next_tid(self, txn):
+        return await self._stmt_next_tid.fetchval()
+
+    @restart_conn_on_exception
     async def get_current_tid(self, txn):
-        async with self.lock:
-            # again, use storage lock here instead of trns lock
-            return await self._stmt_max_tid.fetchval()
+        return await self._stmt_max_tid.fetchval()
 
     async def get_one_row(self, txn, sql, *args, prepare=False):
         conn = await txn.get_connection()
@@ -792,16 +797,16 @@ WHERE tc.constraint_name = '{}_parent_id_id_key' AND tc.constraint_type = 'UNIQU
             txn._db_conn._con._top_xact = None
         return txn._db_conn.transaction(readonly=txn._manager._storage._read_only)
 
+    @restart_conn_on_exception
+    async def _async_db_transaction_factory(self, txn):
+        return self._db_transaction_factory(txn)
+
     async def start_transaction(self, txn, retries=0):
         error = None
         conn = await txn.get_connection()
         async with txn._lock:
-            try:
-                txn._db_txn = self._db_transaction_factory(txn)
-            except asyncpg.exceptions.InterfaceError as ex:
-                async with self.lock:
-                    await self._check_bad_connection(ex)
-                raise
+            txn._db_txn = await self._async_db_transaction_factory(txn)
+
             try:
                 await txn._db_txn.start()
                 return