+++ guillotina/db/interfaces.py
@@ -1,3 +1,4 @@
+from guillotina.interfaces import ICatalogDataAdapter
 from guillotina.interfaces import IDatabase
 from zope.interface import Interface
 
@@ -229,3 +230,8 @@ class IDatabaseManager(Interface):
         '''
         Return whether a db exists or not
         '''
+
+
+class IJSONDBSerializer(ICatalogDataAdapter):
+    '''
+    '''